# src/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Multi-Chain Smart Contract Analytics API
  description: |
    Comprehensive blockchain analytics platform that analyzes smart contracts across multiple chains 
    with intelligent chain isolation and enterprise-grade reliability.
    
    ## Features
    - **Multi-Chain Support**: Ethereum, Lisk, Starknet, Polygon, Base, Arbitrum, Optimism
    - **OAuth Authentication**: Google and GitHub OAuth integration
    - **Multiple Contract Configurations**: Save and manage multiple analysis configurations
    - **Analysis History**: Track all your analyses with detailed results
    - **Comprehensive Analytics**: 20+ DeFi metrics across user lifecycle, activity, financial, and performance categories
    - **Multiple Output Formats**: JSON, CSV, and Markdown reports
    
    ## Authentication
    This API supports two authentication methods:
    1. **JWT Bearer Token**: Use the `/api/auth/login` endpoint to get a token
    2. **API Key**: Use the `X-API-Key` header with your personal API key
    
    ## Rate Limits
    - General endpoints: 100 requests per 15 minutes
    - Analysis endpoints: 10 requests per hour
    - Limits vary by user tier (Free, Pro, Enterprise)
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@analytics-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.analytics-platform.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        avatar:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        tier:
          type: string
          enum: [free, pro, enterprise]
          example: "free"
        oauthProvider:
          type: string
          enum: [google, github]
          nullable: true
          example: "google"
        isActive:
          type: boolean
          example: true
        lastLogin:
          type: string
          format: date-time
          nullable: true
        preferences:
          type: object
          properties:
            notifications:
              type: object
              properties:
                email:
                  type: boolean
                analysis:
                  type: boolean
            defaultChain:
              type: string
              example: "ethereum"
        usage:
          type: object
          properties:
            analysisCount:
              type: integer
              example: 5
            monthlyAnalysisCount:
              type: integer
              example: 3
            lastAnalysis:
              type: string
              format: date-time
              nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: string
          nullable: true
...
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      description: Create a new user account with email and password
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  minLength: 6
                  example: "securepassword123"
                name:
                  type: string
                  example: "John Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  verificationToken:
                    type: string
                    example: "abc123def456"
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      summary: Login user
      description: Authenticate user with email and password
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/oauth/google:
    post:
      summary: Google OAuth login
      description: Authenticate user with Google OAuth access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accessToken
              properties:
                accessToken:
                  type: string
                  example: "ya29.a0AfH6SMC..."
      responses:
        '200':
          description: OAuth login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OAuth login successful"
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/oauth/github:
    post:
      summary: GitHub OAuth login
      description: Authenticate user with GitHub OAuth access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accessToken
              properties:
                accessToken:
                  type: string
                  example: "gho_16C7e42F292c6912E7710c838347Ae178B4a"
      responses:
        '200':
          description: OAuth login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OAuth login successful"
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...